+++ app/lecture/[id]/components/comments.tsx
@@ -1,6 +1,6 @@
 import { HStack, VStack } from '@/components/basic/stack';
 import Button from '@/components/basic/button';
-import CommentView from './comment';
+import CommentView from '@/components/view/CommentView';
 import Link from 'next/link';
 import { useState } from 'react';
 import Select from '@/components/basic/select';
